Refactor: component separation
