+++ src/components/MainPage/MainPage.jsx
@@ -5,15 +5,9 @@ import 'slick-carousel/slick/slick-theme.css';
 import styled from 'styled-components';
 import { wine } from '../../data';
 import CardFlip from './CardFlip';
-import { useNavigate } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-import { LoginState, LoginUserName } from '../../states/LoginState';
+import LoginPermission from './LoginPermission';
 
 function MainPage() {
-  const navigate = useNavigate();
-  const isLoggedIn = useRecoilValue(LoginState);
-  const loginUserName = useRecoilValue(LoginUserName);
-
   const category = [
     { title: '이달의 와인 TOP5', value: 'mention' },
     { title: '선물하기 좋은 와인 TOP5', value: 'present' },
@@ -31,24 +25,7 @@ function MainPage() {
 
   return (
     <MainContainer>
-      <div className='loginWrapper'>
-        {isLoggedIn
-          ? (
-            <>
-              <div><span>{loginUserName}</span>님의 취향 저격 와인<br />AI가 골라드려요!</div>
-              <button onClick={() => navigate('/signin')}>와인 추천받기</button>
-            </>
-          )
-
-          : (
-            <>
-              <div>로그인하고 나만을 위해 준비된<br />와인을 만나보세요!</div>
-              <button onClick={() => navigate('/signin')}>로그인</button>
-            </>
-          )
-        }
-      </div>
-
+      <LoginPermission/>
       <SlideContainer>
         {
           category.map((c, idx) => {
@@ -76,33 +53,6 @@ export default MainPage;
 const MainContainer = styled.div`
   margin: 0 24px;
   span {font-weight: bold}
-  
-  .loginWrapper {
-    height: 66px;
-    border-radius: 25px;
-    background-color: #C57AEA;
-    box-shadow: 4px 8px 10px 0 rgba(0, 0, 0, 0.2);
-    padding: 15px 20px 25px;
-    div {
-      color: #FFFFFF;
-    }
-    button {
-      border-radius: 15px;
-      background-color: #FFFFFF;
-      color: #000000;
-      float: right;
-      border: none;
-      outline: none;
-      cursor: pointer;
-      padding: 8px 22px;
-      font-weight: bold;
-    }
-  }
-  .loginWrapper::after {
-    content: '';
-    display: block;
-    clear: both;
-  }
 `;
 
 const SlideContainer = styled.div`
